Added users and a more app-like UI
I'm using Firebase for auth because it made things really simple, sessions use gorilla sessions as the Echo docs had referenced. All of the sql queries for everything had to be scoped to a user, even for things like just deleting something by id, so someone can't just iterate all numbers and delete everyone's data, it'll only work if they own the row. Most API routes require auth, anything that's not just public at least.

I rewrote some things about the front end. I had to conditionally show a landing page, and it was easier to bake it into the main entry page than interact with the backend. If the call to series hits a 401 it will show the landing page instead of the app. The app now by default shows a non-editable view and will poll to update every 10 minutes for volumes and 60 minutes for series. As a precaution when entering edit mode for a volume (to change pages or notes) it will refresh the volumes. All of this is done so the page can remain open without risking overwriting data if you had updated data from another tab/device. We can always update the UI from polling if it's not editable, and if we load right before editing we know we aren't going to overwrite unsaved data.

I'm deploying this somewhere new so I just edited the base migration file for the new data.

+++ src/server/repositories/readinghistory.go
@@ -12,6 +12,7 @@ type ReadingHistoryEntity struct {
 	VolumeId    int       `db:"volume_id"`
 	CurrentPage int       `db:"current_page"`
 	CreatedAt   time.Time `db:"created_at"`
+	UserId      int       `db:"user_id"`
 }
 
 type ReadingHistoryRepository struct {
@@ -22,20 +23,20 @@ func NewReadingHistoryRepository() *ReadingHistoryRepository {
 	return &ReadingHistoryRepository{db.Connection}
 }
 
-func (r *ReadingHistoryRepository) List(volumeId int) ([]ReadingHistoryEntity, error) {
+func (r *ReadingHistoryRepository) List(userId, volumeId int) ([]ReadingHistoryEntity, error) {
 	history := []ReadingHistoryEntity{}
 
-	err := r.db.Select(&history, "select * from reading_history where volume_id=$1 order by created_at desc", volumeId)
+	err := r.db.Select(&history, "select * from reading_history where volume_id=$1 and user_id=$2 order by created_at desc", volumeId, userId)
 
 	return history, err
 }
 
-func (r *ReadingHistoryRepository) Add(volumeId, currentPage int) error {
-	_, err := r.db.Exec("insert into reading_history (volume_id, current_page, created_at) values ($1, $2, $3)", volumeId, currentPage, time.Now())
+func (r *ReadingHistoryRepository) Add(userId, volumeId, currentPage int) error {
+	_, err := r.db.Exec("insert into reading_history (volume_id, current_page, created_at, user_id) values ($1, $2, $3, $4)", volumeId, currentPage, time.Now(), userId)
 	return err
 }
 
-func (r *ReadingHistoryRepository) Delete(id int) error {
-	_, err := r.db.Exec("delete from reading_history where id=$1", id)
+func (r *ReadingHistoryRepository) Delete(userId, id int) error {
+	_, err := r.db.Exec("delete from reading_history where id=$1 and user_id=$2", id, userId)
 	return err
 }
