+++ src/server/repositories/volume.go
@@ -11,6 +11,7 @@ type VolumeEntity struct {
 	Id          int       `db:"id"`
 	Name        string    `db:"name"`
 	SeriesId    string    `db:"series_id"`
+	UserId      string    `db:"user_id"`
 	Notes       string    `db:"notes"`
 	CurrentPage int       `db:"current_page"`
 	CreatedAt   time.Time `db:"created_at"`
@@ -20,33 +21,33 @@ type VolumeRepository struct {
 	db *sqlx.DB
 }
 
-func (v *VolumeRepository) FindOne(volumeId int) (VolumeEntity, error) {
-	volume := VolumeEntity{}
-
-	err := v.db.Get(&volume, "select * from volumes where id=$1", volumeId)
-
-	return volume, err
-}
-
 func NewVolumeRepository() *VolumeRepository {
 	return &VolumeRepository{db.Connection}
 }
 
-func (v *VolumeRepository) List(seriesId int) ([]VolumeEntity, error) {
+func (v *VolumeRepository) FindOne(userId, volumeId int) (VolumeEntity, error) {
+	volume := VolumeEntity{}
+
+	err := v.db.Get(&volume, "select * from volumes where id=$1 and user_id=$2", volumeId, userId)
+
+	return volume, err
+}
+
+func (v *VolumeRepository) List(userId, seriesId int) ([]VolumeEntity, error) {
 	volumes := []VolumeEntity{}
 
-	err := v.db.Select(&volumes, "select * from volumes where series_id=$1 order by created_at asc", seriesId)
+	err := v.db.Select(&volumes, "select * from volumes where series_id=$1 and user_id=$2 order by created_at asc", seriesId, userId)
 
 	return volumes, err
 }
 
-func (v *VolumeRepository) Add(seriesId int, name string) error {
-	_, err := v.db.Exec("insert into volumes (series_id, name, created_at) values ($1, $2, $3)", seriesId, name, time.Now())
+func (v *VolumeRepository) Add(userId, seriesId int, name string) error {
+	_, err := v.db.Exec("insert into volumes (series_id, name, created_at, user_id) values ($1, $2, $3, $4)", seriesId, name, time.Now(), userId)
 	return err
 }
 
-func (v *VolumeRepository) Delete(volumeId int) error {
-	_, err := v.db.Exec("delete from volumes where id=$1", volumeId)
+func (v *VolumeRepository) Delete(userId, volumeId int) error {
+	_, err := v.db.Exec("delete from volumes where id=$1 and user_id=$2", volumeId, userId)
 
 	return err
 }
@@ -57,7 +58,7 @@ type VolumeEntityUpdateArgs struct {
 	CurrentPage int    `db:"current_page"`
 }
 
-func (v *VolumeRepository) Update(volumeId int, update *VolumeEntityUpdateArgs) error {
-	_, err := v.db.Exec("update volumes set notes=$1, current_page=$2, name=$3 where id=$4", update.Notes, update.CurrentPage, update.Name, volumeId)
+func (v *VolumeRepository) Update(userId, volumeId int, update *VolumeEntityUpdateArgs) error {
+	_, err := v.db.Exec("update volumes set notes=$1, current_page=$2, name=$3 where id=$4 and user_id=$5", update.Notes, update.CurrentPage, update.Name, volumeId, userId)
 	return err
 }
